\name{print.mevd}
\alias{print.mevd}
\title{Print output of an mevd object.}
\usage{
     \method{print}{mevd}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
     \item{x}{mevd object.}
     
     \item{digits}{Number of digits.}
     \item{...}{Additional parameters.}
}
\description{
     Print output of a mevd object.
} 